Saving a checkpoint into a branch for GoLang core library. Some wins, but going to look at another approach.
Wins contained here:
 - Nice package structure. One module, several packages. Confirmed external projects can import select packages.
 - Build tools that produce an xcframework
 - Go mobile works great, so should compile to Android
 - confirmed iOS 11 SDK support
 - Confirmed integrating into an iOS project
 - Confirmed integrating a package from core project into separate project (cswasm folder)
 - Primary folder is “go”, other “go” prefix folders were earlier attempts.

The issue is: the binary in the xcframework is 2MB (with flags), and the wasm file is 2MB, and that’s just for a basic hello world app. As we add libraries it will grow. This is known, and optimizations will help a bit, but not a ton. This is too big to deploy on Cloudflare free workers, and may even grow to be too big for paid workers. It’s never viable for browser WASM. It’s also a lot of binary weight even on a mobile app. They do compress well, so not a deal breaker but would need to start . Going to investigate another option, and maybe come back to this.